requirements.txt
Flask==2.0.1
Werkzeug==2.0.3
PyYAML==6.0
gunicorn==20.1.0
